fix missing METAR data with repetitive downloads
Bug report:
If a station did not yield METAR data for certain dates, nothing would be stored in the database.
This in turn resulted in another request for these stations and dates during the next query.

Fix:
Check missing data after download and write a single entry for each day in the database with METAR = None.

Additional:
- if data of station is available for complete timespan, don't download